Commented code was in documentation; Config.py was removed and instructions was in the readme.md for user to create their own config.py Images are in the static/img Flask is in the app directory now
