+++ app/Flask/requirements.txt
@@ -1,3 +1,4 @@
+numpy==1.14.1
 Flask==0.10.1
 Flask-SQLAlchemy==2.0
 Flask-WTF==0.10.3
