+++ app/Flask/static/script.js
