+++ app/Flask/templates/index.html
@@ -13,13 +13,13 @@
         <title>FIFA18_Prediction</title>
     </head>
     <body>
-        <header style="background: linear-gradient(rgba(0, 0, 0, 0.7), rgba(0, 0, 0, 0.7)), url({{url_for("static", filename="FIFA18_background.jpg")}}); background-size: 100%; background-position: center; height: 100vh; background-attachment: fixed;">
+        <header style="background: linear-gradient(rgba(0, 0, 0, 0.7), rgba(0, 0, 0, 0.7)), url({{url_for("static", filename="img/FIFA18_background.jpg")}}); background-size: 100%; background-position: center; height: 100vh; background-attachment: fixed;">
 
             
             <nav>
                 <div class="row">
-                    <img src="{{ url_for('static', filename = 'fifa18_logo.png') }}" alt="FIFA logo" class="logo">
-                    <img src="{{ url_for('static', filename = 'fifa18_logo.png') }}" alt="logo" class="logo-black">
+                    <img src="{{ url_for('static', filename = 'img/fifa18_logo.png') }}" alt="FIFA logo" class="logo">
+                    <img src="{{ url_for('static', filename = 'img/fifa18_logo.png') }}" alt="logo" class="logo-black">
                     <ul class="main-nav js--main-nav">
                         <li><a href="https://www.easports.com/fifa" target="_blank">about us</a></li>
                         <li><a href="https://www.kaggle.com/thec03u5/fifa-18-demo-player-dataset" target="_blank">Model</a></li>
